implemented FeedParser class and added sample program
